+++ src/test-acceptance/resources/jscover/ui/scripts/script-c.js
@@ -1,0 +1,16 @@
+(function () {
+    function increment(number) {
+        return number + 1;
+    }
+
+    function sign(number) {
+        if (number < 0)
+            return -1;
+        else if (number > 0)
+            return 1;
+        else
+            return 0;
+    }
+    increment(1);
+    sign(-1);
+}());