+++ History.md
@@ -1,8 +1,10 @@
 1.0.7 / 2014-??-??
 ==================
   * Add sort by file name in the coverage report web UI (https://github.com/tntim96/JSCover/issues/108)
+  * Fix handling of 'N/A' when sorting by line coverage (https://github.com/tntim96/JSCover/issues/115)
   * Internal: Upgrade selenium to 2.39
               Use tntim96 Maven repository version of Rhino
+              Add more UI tests
 
 1.0.6 / 2013-11-03
 ==================